pyproject.toml
[project]
name = "folderstructure"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "robotframework>=7.2",
    "robotframework-browser>=19.1.2",
]

[dependency-groups]
dev = [
    "robotframework-tidy>=4.16.0",
    "ruff>=0.9.2",
]
